Bug Fixes
Former-commit-id: f3a2179fa340dd16d24ff0383f51a9ae0d4a87a5

+++ src/librapid/config.hpp
@@ -157,7 +157,7 @@ namespace librapid
 	// #endif // LIBRAPID_HAS_OPENBLAS
 	// }
 
-	inline void setNumThreads(int num)
+	inline void setNumThreads(int64_t num)
 	{
 	#if defined(LIBRAPID_HAS_OPENBLAS)
 		openblas_set_num_threads(num);
